{{cookiecutter.repo_name}}/requirements.txt
Django==2.2.4
psycopg2-binary==2.7.7
uWSGI==2.0.15
djangorestframework==3.9.1
django-storages==1.6.5
django-ses==0.8.4
Jinja2==2.10.1
setuptools==38.2.4
raven==6.4.0
zappa==0.46.2
